Checkout and Publish do the file saves they need. No custom script yet from Chris
