+++ nuke-tools/init.py
@@ -1,6 +1,5 @@
 import nuke
 
 # Add subfolders to path to make this a little cleaner
-nuke.pluginAddPath( './gizmos' )
 nuke.pluginAddPath('./icons')
-
+nuke.pluginAddPath('./scripts')
